Move memory utilities to their own folder.
