+++ GridKit/LinearAlgebra/CMakeLists.txt
@@ -1,7 +1,3 @@
-add_subdirectory(cpu)
-
 add_subdirectory(SparseMatrix)
 add_subdirectory(DenseMatrix)
 add_subdirectory(Vector)
-
-install(FILES ${CMAKE_CURRENT_SOURCE_DIR}/MemoryUtils.hpp DESTINATION include/GridKit/LinearAlgebra)
