+++ GridKit/MemoryUtilities/MemoryUtils.tpp
